refactor(jtms): migrate JTMSSession/ExtendedBelief imports to canonical services/jtms/ (Issue #263)
The classes ExtendedBelief and JTMSSession were extracted to
services/jtms/extended_belief.py but 4 files still imported them from
the stale agents/jtms_agent_base.py location. This creates confusion
about the canonical import path and risks divergence.

Migration:
- agents/watson_jtms/agent.py: JTMSSession → services.jtms
- agents/sherlock_jtms_agent.py: ExtendedBelief → services.jtms
- agents/jtms_communication_hub.py: JTMSSession → services.jtms
- core/state_manager_plugin.py: JTMSSession → services.jtms (4 sites)

JTMSAgentBase remains in agents/jtms_agent_base.py (not yet extracted).
