+++ argumentation_analysis/agents/watson_jtms/agent.py
@@ -1,4 +1,5 @@
-from argumentation_analysis.agents.jtms_agent_base import JTMSSession, JTMSAgentBase
+from argumentation_analysis.services.jtms import JTMSSession
+from argumentation_analysis.agents.jtms_agent_base import JTMSAgentBase
 from .consistency import ConsistencyChecker
 from .validation import FormalValidator
 from .critique import CritiqueEngine
