+++ argumentation_analysis/core/state_manager_plugin.py
@@ -337,7 +337,7 @@ class StateManagerPlugin:
         )
         try:
             # Import here to avoid circular dependency
-            from argumentation_analysis.agents.jtms_agent_base import JTMSSession
+            from argumentation_analysis.services.jtms import JTMSSession
             import json
 
             # Get or create JTMS session from state
@@ -387,7 +387,7 @@ class StateManagerPlugin:
             f"Appel jtms_add_justification: IN={len(in_list)}, OUT={len(out_list)}, conclusion={conclusion}"
         )
         try:
-            from argumentation_analysis.agents.jtms_agent_base import JTMSSession
+            from argumentation_analysis.services.jtms import JTMSSession
 
             # Get or create JTMS session
             if not hasattr(self._state, "_jtms_session"):
@@ -415,7 +415,7 @@ class StateManagerPlugin:
         """Query JTMS beliefs with ExtendedBelief metadata via StateManagerPlugin (#214)."""
         self._logger.info(f"Appel jtms_query_beliefs: filter={agent_filter}")
         try:
-            from argumentation_analysis.agents.jtms_agent_base import JTMSSession
+            from argumentation_analysis.services.jtms import JTMSSession
             import json
 
             # Get JTMS session
@@ -458,7 +458,7 @@ class StateManagerPlugin:
         """Check JTMS consistency and return conflict report via StateManagerPlugin (#214)."""
         self._logger.info("Appel jtms_check_consistency")
         try:
-            from argumentation_analysis.agents.jtms_agent_base import JTMSSession
+            from argumentation_analysis.services.jtms import JTMSSession
             import json
 
             # Get JTMS session
